REFAC: Refactor derectory and code responsibilities
This change forcus to make it easy to maintain
