+++ src/config/mod.rs
@@ -1,0 +1,5 @@
+pub mod paths;
+pub mod user;
+
+pub use paths::{config_dir, export_dir};
+pub use user::{UserConfig, load, save};
