+++ src/hyprland/models.rs
@@ -1,4 +1,4 @@
-use crate::icons::get_icon;
+use crate::ui::styling::icons::get_icon;
 use serde::{Deserialize, Serialize};
 
 /// Options for searching keybindings
@@ -74,18 +74,6 @@ impl KeyBindings {
         self.entries.push(entry);
     }
 
-    /// Filter entries by search query
-    pub fn filter(&self, query: &str, options: &SearchOptions) -> Vec<&KeyBindEntry> {
-        if query.is_empty() {
-            self.entries.iter().collect()
-        } else {
-            self.entries
-                .iter()
-                .filter(|e| e.matches(query, options))
-                .collect()
-        }
-    }
-
     /// Export as JSON
     pub fn to_json(&self) -> Result<String, serde_json::Error> {
         serde_json::to_string_pretty(self)
@@ -127,61 +115,5 @@ impl KeyBindings {
 impl Default for KeyBindings {
     fn default() -> Self {
         Self::new()
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_to_dmenu() {
-        // 1. No modifier, with description
-        let entry1 = KeyBindEntry::new(
-            String::new(),
-            "Return".to_string(),
-            "exec kitty".to_string(),
-            "Terminal".to_string(),
-        );
-        // 2. With modifiers, with description
-        let entry2 = KeyBindEntry::new(
-            "SUPER+SHIFT".to_string(),
-            "Q".to_string(),
-            "killactive".to_string(),
-            "Kill window".to_string(),
-        );
-        // 3. With modifiers, no description
-        let entry3 = KeyBindEntry::new(
-            "CTRL+ALT".to_string(),
-            "F1".to_string(),
-            "exec firefox".to_string(),
-            String::new(),
-        );
-        // 4. With modifiers, no description, no command
-        let entry4 = KeyBindEntry::new(
-            "CTRL+ALT".to_string(),
-            "F2".to_string(),
-            String::new(),
-            String::new(),
-        );
-
-        let kb = KeyBindings {
-            entries: vec![entry1, entry2, entry3, entry4],
-        };
-
-        let dmenu = kb.to_dmenu();
-        let lines: Vec<&str> = dmenu.lines().collect();
-
-        // 1. No modifier, icon only
-        assert_eq!(lines[0], "󰌑 : Terminal");
-
-        // 2. Modifiers, icons
-        assert_eq!(lines[1], " +  󰘶  + Q : Kill window");
-
-        // 3. Modifiers, fallback to key text if not in icon table
-        assert_eq!(lines[2], "CTRL + ALT + F1 : exec firefox");
-
-        // 4. Modifiers, no description, no command
-        assert_eq!(lines[3], "CTRL + ALT + F2 : ");
     }
 }
