+++ src/tests/mod.rs
@@ -1,0 +1,4 @@
+mod icons;
+mod models;
+mod parser;
+mod table;
