+++ src/ui/mod.rs
@@ -1,5 +1,6 @@
 pub mod header;
 pub mod options;
+pub mod styling;
 pub mod table;
 pub mod types;
 pub mod zen;
