+++ src/ui/styling/icons.rs
@@ -48,46 +48,3 @@ pub fn get_icon(key: &str) -> String {
     // Fallback: return the key itself
     key.to_string()
 }
-
-#[cfg(test)]
-mod tests {
-    use super::get_icon;
-
-    #[test]
-    fn test_get_icon() {
-        let cases: [(&str, &str); 28] = [
-            ("SUPER", ""),
-            ("SHIFT", " 󰘶 "),
-            ("RETURN", "󰌑"),
-            ("ENTER", "󰌑"),
-            ("SEMICOLON", ";"),
-            ("DELETE", "DEL"),
-            ("TAB", "TAB"),
-            ("LEFT", "󰜱"),
-            ("RIGHT", "󰜴"),
-            ("UP", "󰜷"),
-            ("DOWN", "󰜮"),
-            ("mouse_down", "󱕐"),
-            ("mouse_up", "󱕑"),
-            ("mouse:272", "󰍽"),
-            ("mouse:273", "󰍽"),
-            ("XF86AudioRaiseVolume", ""),
-            ("XF86AudioLowerVolume", ""),
-            ("XF86AudioMute", ""),
-            ("XF86AudioMicMute", "󰍭"),
-            ("XF86MonBrightnessUp", "󰃠"),
-            ("XF86MonBrightnessDown", "󰃞"),
-            ("XF86AudioNext", "󰙡"),
-            ("XF86AudioPause", ""),
-            ("XF86AudioPlay", ""),
-            ("XF86AudioPrev", "󰙣"),
-            ("UNKNOWN_KEY", "UNKNOWN_KEY"),
-            ("A", "A"),
-            ("123", "123"),
-        ];
-
-        for (input, expected) in &cases {
-            assert_eq!(get_icon(input), *expected);
-        }
-    }
-}
