+++ src/ui/table.rs
@@ -1,6 +1,6 @@
 use super::types::{ColumnVisibility, SortColumn, SortState};
-use crate::icons::get_icon;
-use crate::models::KeyBindEntry;
+use crate::hyprland::KeyBindEntry;
+use crate::ui::styling::icons::get_icon;
 use eframe::egui;
 use egui_extras::{Column, TableBuilder};
 
@@ -58,7 +58,7 @@ fn render_header_cell(
     clicked
 }
 
-fn is_nerd_font_icon(text: &str) -> bool {
+pub(crate) fn is_nerd_font_icon(text: &str) -> bool {
     // Check if text contains Nerd Font Unicode characters (Private Use Areas)
     text.chars().any(|c| {
         let code = c as u32;
@@ -323,35 +323,4 @@ pub fn render_table(
         });
 
     clicked_column
-}
-
-#[cfg(test)]
-mod tests {
-
-    use super::is_nerd_font_icon;
-
-    #[test]
-    fn test_is_nerd_font_icon() {
-        // Test with NerdFonts
-        let nerd_fonts: [&str; 21] = [
-            "", "󰘶", "󰌑", "󰜱", "󰜴", "󰜷", "󰜮", "󱕐", "󱕑", "󰍽", "󰍽", "", "", "", "󰍭", "󰃠", "󰃞",
-            "󰙡", "", "", "󰙣",
-        ];
-
-        let non_nerd_fonts: [&str; 5] = [";", "A", "DEL", "TAB", "1"];
-
-        for icon in &nerd_fonts {
-            assert!(
-                is_nerd_font_icon(icon),
-                "Expected '{icon}' to be identified as a Nerd Font icon"
-            );
-        }
-
-        for text in &non_nerd_fonts {
-            assert!(
-                !is_nerd_font_icon(text),
-                "Expected '{text}' to NOT be identified as a Nerd Font icon"
-            );
-        }
-    }
 }
